backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pymysql==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
Pillow==10.3.0
pypdf==4.2.0
reportlab==4.2.0
aiofiles==23.2.1
python-dotenv==1.0.1
cryptography==42.0.7
